Moved many old files to appropriate 'deprecated' namespace
