--- Project-DOOM/Assets/Core/_Scripts/Enemy/BillboardJob.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c47024f0d2745044957d96a20703b17
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
