+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Enemy DOTS/V2/NewEnemy.cs
@@ -3,19 +3,22 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class NewEnemy : MonoBehaviour
+namespace ProjectDOOM.deprecated.V2
 {
-	private SpriteRenderer _spriteRendererComponent;
-
-	private void Awake()
+	public class NewEnemy : MonoBehaviour
 	{
-		_spriteRendererComponent = GetComponentInChildren<SpriteRenderer>();
-	}
+		private SpriteRenderer _spriteRendererComponent;
 
-	public void ApplyRotationAndIndex(Quaternion rotation, int index)
-	{
-		// Implement how you update the rotation and index for an enemy.
-		_spriteRendererComponent.transform.rotation = rotation;
-		// You can use the index as needed.
+		private void Awake()
+		{
+			_spriteRendererComponent = GetComponentInChildren<SpriteRenderer>();
+		}
+
+		public void ApplyRotationAndIndex(Quaternion rotation, int index)
+		{
+			// Implement how you update the rotation and index for an enemy.
+			_spriteRendererComponent.transform.rotation = rotation;
+			// You can use the index as needed.
+		}
 	}
 }
