+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Enemy DOTS/V4/EnemyJobDOTS.cs
@@ -7,106 +7,109 @@ using Unity.Jobs;
 using Unity.Mathematics;
 using UnityEngine;
 
-[BurstCompile]
-public struct EnemyJobDOTS : IJobParallelFor
+namespace ProjectDOOM.deprecated.V4
 {
-	[ReadOnly]
-	public float3 cameraPosition;
-	[ReadOnly]
-	public float cameraRotationY;
-	
-	[ReadOnly]
-	public NativeArray<float3> enemyPositions;
-	[ReadOnly]
-	public NativeArray<float3> enemyForwardDirections;
-	[ReadOnly]
-	public NativeArray<float3> anglesArray;
-	
-	[NativeDisableParallelForRestriction]
-	public NativeArray<int> returnedRotationIndex;
-	[NativeDisableParallelForRestriction]
-	public NativeArray<Quaternion> returnedRotation;
-	
-	public void Execute(int index)
+	[BurstCompile]
+	public struct EnemyJobDOTS : IJobParallelFor
 	{
-		returnedRotationIndex[index] = GetRotationIndex(
-			BetterDetermineAngle(
-				enemyPositions[index],
-				cameraPosition,
-				enemyForwardDirections[index]
+		[ReadOnly]
+		public float3 cameraPosition;
+		[ReadOnly]
+		public float cameraRotationY;
+	
+		[ReadOnly]
+		public NativeArray<float3> enemyPositions;
+		[ReadOnly]
+		public NativeArray<float3> enemyForwardDirections;
+		[ReadOnly]
+		public NativeArray<float3> anglesArray;
+	
+		[NativeDisableParallelForRestriction]
+		public NativeArray<int> returnedRotationIndex;
+		[NativeDisableParallelForRestriction]
+		public NativeArray<Quaternion> returnedRotation;
+	
+		public void Execute(int index)
+		{
+			returnedRotationIndex[index] = GetRotationIndex(
+				BetterDetermineAngle(
+					enemyPositions[index],
+					cameraPosition,
+					enemyForwardDirections[index]
 				),
-			anglesArray
+				anglesArray
 			);
 
-		returnedRotation[index] = Quaternion.Euler(0f, cameraRotationY, 0f);
-	}
+			returnedRotation[index] = Quaternion.Euler(0f, cameraRotationY, 0f);
+		}
 	
-	[BurstCompile]
-	private int GetRotationIndex(float currentAngle, NativeArray<float3> angles)
-	{
-		for (int i = 0; i < angles.Length; i++)
+		[BurstCompile]
+		private int GetRotationIndex(float currentAngle, NativeArray<float3> angles)
 		{
-			// x = min, y = max, z = index
-			if (angles[i].z == 0f) // special case for front
+			for (int i = 0; i < angles.Length; i++)
 			{
-				if (currentAngle > angles[i].x && currentAngle < angles[i].y)
+				// x = min, y = max, z = index
+				if (angles[i].z == 0f) // special case for front
+				{
+					if (currentAngle > angles[i].x && currentAngle < angles[i].y)
+					{
+						return (int) angles[i].z;
+					}
+				}
+				else if (angles[i].z == 4f) // special case for back
+				{
+					if (currentAngle <= angles[i].x || currentAngle >= angles[i].y)
+					{
+						return (int) angles[i].z;
+					}
+				}
+				else // everything else
+				if (currentAngle >= angles[i].x && currentAngle < angles[i].y)
 				{
 					return (int) angles[i].z;
 				}
 			}
-			else if (angles[i].z == 4f) // special case for back
-			{
-				if (currentAngle <= angles[i].x || currentAngle >= angles[i].y)
-				{
-					return (int) angles[i].z;
-				}
-			}
-			else // everything else
-			if (currentAngle >= angles[i].x && currentAngle < angles[i].y)
-			{
-				return (int) angles[i].z;
-			}
+
+			return 0;
 		}
 
-		return 0;
-	}
+		[BurstCompile]
+		private float BetterDetermineAngle(float3 a, float3 b, float3 aDir)
+		{
+			float3 targetPosition = new float3(b.x, a.y, b.z);
+			float3 targetDirection = targetPosition - a;
 
-	[BurstCompile]
-	private float BetterDetermineAngle(float3 a, float3 b, float3 aDir)
-	{
-		float3 targetPosition = new float3(b.x, a.y, b.z);
-		float3 targetDirection = targetPosition - a;
+			return SignedAngle(targetDirection, aDir, new float3(0f, 1f, 0f));
+		}
 
-		return SignedAngle(targetDirection, aDir, new float3(0f, 1f, 0f));
-	}
+		#region Calculating a better SignedAngle
 
-	#region Calculating a better SignedAngle
-
-	[BurstCompile]
-	private float Angle(float3 from, float3 to)
-	{
-		// float num = (float) Math.Sqrt((double) from.sqrMagnitude * (double) to.sqrMagnitude);
-		float num = math.sqrt(math.lengthsq(from) * math.lengthsq(to));
-		// return num < 1.0000000036274937E-15 ? 0.0f : (float) Math.Acos((double) Mathf.Clamp(Vector3.Dot(from, to) / num, -1f, 1f)) * 57.29578f;
-		return num < 1.0000000036274937E-15 ? 0.0f : math.acos(math.clamp(math.dot(from, to) / num, -1f, 1f)) * 57.29578f;
-	}
+		[BurstCompile]
+		private float Angle(float3 from, float3 to)
+		{
+			// float num = (float) Math.Sqrt((double) from.sqrMagnitude * (double) to.sqrMagnitude);
+			float num = math.sqrt(math.lengthsq(from) * math.lengthsq(to));
+			// return num < 1.0000000036274937E-15 ? 0.0f : (float) Math.Acos((double) Mathf.Clamp(Vector3.Dot(from, to) / num, -1f, 1f)) * 57.29578f;
+			return num < 1.0000000036274937E-15 ? 0.0f : math.acos(math.clamp(math.dot(from, to) / num, -1f, 1f)) * 57.29578f;
+		}
 	
-	[BurstCompile]
-	private float SignedAngle(float3 from, float3 to, float3 axis)
-	{
-		float num1 = Angle(from, to);
-		float num2 = (from.y * to.z - from.z * to.y);
-		float num3 = (from.z * to.x - from.x * to.z);
-		float num4 = (from.x * to.y - from.y * to.x);
-		float num5 = Sign(axis.x * num2 + axis.y * num3 + axis.z * num4);
-		return num1 * num5;
-	}
+		[BurstCompile]
+		private float SignedAngle(float3 from, float3 to, float3 axis)
+		{
+			float num1 = Angle(from, to);
+			float num2 = (from.y * to.z - from.z * to.y);
+			float num3 = (from.z * to.x - from.x * to.z);
+			float num4 = (from.x * to.y - from.y * to.x);
+			float num5 = Sign(axis.x * num2 + axis.y * num3 + axis.z * num4);
+			return num1 * num5;
+		}
 	
-	[BurstCompile]
-	private float Sign(float f)
-	{
-		return f >= 0.0 ? 1f : -1f;
-	}
+		[BurstCompile]
+		private float Sign(float f)
+		{
+			return f >= 0.0 ? 1f : -1f;
+		}
 
-	#endregion
+		#endregion
+	}
 }
