--- Project-DOOM/Assets/Core/_Scripts/Enemy/Enemy DOTS/V5/ISpriteComponentReferences.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface ISpriteComponentReferences
-{
-	public void UpdateAnimatorIndex(int index);
-	
-	public void UpdateSpriteDirection(float yRotation);
-	
-	public Transform GetSpriteTransform();
-	
-	public Transform GetTransform();
-}
