+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Enemy DOTS/V5/SimpleSpriteSwitcher.cs
@@ -6,7 +6,7 @@ using Unity.Jobs;
 using Unity.Mathematics;
 using UnityEngine;
 
-namespace DOTS.V5
+namespace ProjectDOOM.deprecated.DOTS
 {
 	public class SimpleSpriteSwitcher : MonoBehaviour
 	{
