+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Enemy DOTS/V5/SpriteSwitcherJob.cs
@@ -7,7 +7,7 @@ using Unity.Jobs;
 using Unity.Mathematics;
 using UnityEngine;
 
-namespace DOTS.V5
+namespace ProjectDOOM.deprecated.DOTS
 {
 	public struct SpriteSwitcherJob : IJob
 	{
