--- Project-DOOM/Assets/Core/_Scripts/Enemy/EnemyBillboardManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5754dc0cb17e1e4d9b506bc6320ff26
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
