--- Project-DOOM/Assets/Core/_Scripts/Enemy/EnemyComponents.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3d9c154571a4096438523b31f87132e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
