--- Project-DOOM/Assets/Core/_Scripts/Enemy/EnemySpriteManager.cs
@@ -1,27 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemySpriteManager : MonoBehaviour
-{
-	private void OnEnable()
-	{
-		
-	}
-
-	private void OnDisable()
-	{
-		
-	}
-
-	private void Start()
-	{
-		
-	}
-
-	private void Update()
-	{
-		
-	}
-}
