--- Project-DOOM/Assets/Core/_Scripts/Enemy/EnemySpriteManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 17a6f5b54a1ff3e4fb6302c3f3ee1ab0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
