+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Sprite Billboard Handling.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cba346f1f51749b498f81ffa3e71fe34
+guid: dde7396b23f408c458a890bd1fe5cbf7
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
