+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Sprite Billboard Handling/EnemySpriteManager.cs
@@ -6,7 +6,7 @@ using Unity.Jobs;
 using Unity.Mathematics;
 using UnityEngine;
 
-namespace DOTS.V6
+namespace EnemySpriteHandling
 {
 	public class EnemySpriteManager : MonoBehaviour
 	{
