+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Sprite Billboard Handling/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5896dc9d56a465c4ea8ef003a5ac8513
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
