+++ Project-DOOM/Assets/Core/_Scripts/Enemy/Sprite Billboard Handling/Interfaces/ISpriteComponentReferences.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace EnemySpriteHandling
+{
+	public interface ISpriteComponentReferences
+	{
+		public void UpdateAnimatorIndex(int index);
+	
+		public void UpdateSpriteDirection(float yRotation);
+	
+		public Transform GetSpriteTransform();
+	
+		public Transform GetTransform();
+	}	
+}
