--- Project-DOOM/Assets/Core/_Scripts/Enemy/SpriteSelector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60c8b90176af30940ac1ba44150ddc09
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
